# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Lexy AI Hiring Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /prep/generate:
    post:
      operationId: generatePrepPlan
      tags: [prep]
      summary: Generate prep plan for candidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeneratePrepRequest"
      responses:
        "200":
          description: Prep plan generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrepPlan"
  /prep/sessions:
    get:
      operationId: listPrepSessions
      tags: [prep]
      summary: List prep sessions
      parameters:
        - name: candidateId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Prep sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PrepSession"
    post:
      operationId: startPrepSession
      tags: [prep]
      summary: Start prep session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartPrepSessionRequest"
      responses:
        "201":
          description: Prep session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrepSession"
  /prep/sessions/{sessionId}/answer:
    post:
      operationId: submitPrepAnswer
      tags: [prep]
      summary: Submit prep answer
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitAnswerRequest"
      responses:
        "200":
          description: Feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrepFeedback"

  # COORDINATOR
...
components:
  schemas:
...
    SubmitAnswerRequest:
      type: object
      properties:
        questionId:
          type: string
        answerText:
          type: string
      required:
        - questionId
        - answerText
...
    GeneratePrepRequest:
      type: object
      required: [candidateId, jobId]
      properties:
        candidateId:
          type: string
        jobId:
          type: string
        mode:
          type: string
    PrepPlan:
      type: object
      properties:
        id:
          type: string
        candidateId:
          type: string
        jobId:
          type: string
        mode:
          type: string
          enum: [quick, full, mock_interview, behavioral, technical, competency]
        likelyQuestions:
          type: array
          items:
            type: string
        keySkillsToFocus:
          type: array
          items:
            type: string
        preparationTips:
          type: array
          items:
            type: string
        estimatedPrepTimeMinutes:
          type: integer
        readinessScore:
          type: number
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - candidateId
        - jobId
        - mode
        - likelyQuestions
        - keySkillsToFocus
        - preparationTips
        - estimatedPrepTimeMinutes
        - createdAt
    PrepSession:
      type: object
      properties:
        id:
          type: string
        candidateId:
          type: string
        jobId:
          type: string
        mode:
          type: string
        status:
          type: string
          enum: [active, completed]
        questionsAnswered:
          type: integer
        totalQuestions:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - candidateId
        - jobId
        - mode
        - status
        - questionsAnswered
        - totalQuestions
        - createdAt
    StartPrepSessionRequest:
      type: object
      properties:
        jobId:
          type: string
        mode:
          type: string
          enum: [quick, full, mock_interview, behavioral, technical, competency]
      required:
        - jobId
        - mode
    PrepFeedback:
      type: object
      properties:
        score:
          type: number
        feedback:
          type: string
        strengths:
          type: array
          items:
            type: string
        improvements:
          type: array
          items:
            type: string
        sampleAnswer:
          type: string
          nullable: true
        nextQuestion:
          type: string
          nullable: true
        isComplete:
          type: boolean
      required:
        - score
        - feedback
        - strengths
        - improvements
        - isComplete
...